+++ src/app/content/components/component-demos/highlight/demos/highlight-demo.module.ts
@@ -1,14 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { HighlightDemoBasicComponent } from './highlight-demo-basic/highlight-demo-basic.component';
+import { HighlightDemoHtmlComponent } from './highlight-demo-html/highlight-demo-html.component';
+import { HighlightDemoTsComponent } from './highlight-demo-ts/highlight-demo-ts.component';
 import { CovalentHighlightModule } from '@covalent/highlight';
 import { HighlightDemoComponent } from './highlight-demo.component';
 import { HighlightDemoRoutingModule } from './highlight-demo-routing.module';
 import { DemoModule } from '../../../../../components/shared/demo-tools/demo.module';
+import { HighLightDemoCssSharedModule } from './highlight-demo-css/highlight-demo-css-shared';
 
 @NgModule({
-  declarations: [HighlightDemoComponent, HighlightDemoBasicComponent],
+  declarations: [HighlightDemoComponent, HighlightDemoHtmlComponent, HighlightDemoTsComponent],
   imports: [
+    HighLightDemoCssSharedModule,
     DemoModule,
     HighlightDemoRoutingModule,
     /** Covalent Modules */