+++ src/app/services/github.service.ts
@@ -9,6 +9,21 @@ const GITHUB_URL: string = 'https://api.github.com';
 export class GitHubService {
   constructor(private _http: HttpClient) {}
 
+  getVersionDirectories(): Observable<string[]> {
+    return this._http
+      .get<{ name: string; type: string }[]>(GITHUB_URL + '/repos/Teradata/covalent/contents?ref=gh-pages')
+      .pipe(
+        map((data: { name: string; type: string }[]) => {
+          return data
+            .filter((row: { name: string; type: string }) => row.type === 'dir' && row.name.startsWith('v'))
+            .map((row: { name: string; type: string }) => row.name)
+            .sort((a: string, b: string) => {
+              return a < b ? 1 : -1;
+            });
+        }),
+      );
+  }
+
   queryStartCount(): Observable<number> {
     return this._http.get(GITHUB_URL + '/search/repositories?q=repo:Teradata/covalent').pipe(
       map((data: any) => {