+++ src/platform/markdown-navigator/markdown-navigator.component.spec.ts
@@ -7,9 +7,11 @@ import {
   IMarkdownNavigatorCompareWith,
 } from './markdown-navigator.component';
 import { By } from '@angular/platform-browser';
-import { Component, DebugElement } from '@angular/core';
+import { Component, DebugElement, Type } from '@angular/core';
 import { CovalentMarkdownNavigatorModule } from './markdown-navigator.module';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { HttpClientTestingModule, HttpTestingController, TestRequest } from '@angular/common/http/testing';
+import { HttpClient } from '@angular/common/http';
 
 const RAW_MARKDOWN_HEADING: string = 'Heading';
 const RAW_MARKDOWN: string = `# ${RAW_MARKDOWN_HEADING}`;
@@ -27,6 +29,19 @@ const RAW_MARKDOWN_ITEM: IMarkdownNavigatorItem[] = [
 
 const FLAT_MIXED_ITEMS: IMarkdownNavigatorItem[] = [...URL_ITEM, ...RAW_MARKDOWN_ITEM];
 
+const ITEMS_WITH_CUSTOM_ICONS: IMarkdownNavigatorItem[] = [
+  { ...URL_ITEM[0], icon: 'whatshot' },
+  { ...RAW_MARKDOWN_ITEM[0], icon: 'touch_app' },
+];
+
+const ITEMS_WITH_DESCRIPTIONS: IMarkdownNavigatorItem[] = [
+  {
+    ...URL_ITEM[0],
+    description: 'An introduction into what Covalent is all about',
+  },
+  { ...RAW_MARKDOWN_ITEM[0], description: 'An example of using raw markdown' },
+];
+
 const NESTED_MIXED_ITEMS: IMarkdownNavigatorItem[] = [
   {
     title: 'First item',
@@ -35,6 +50,43 @@ const NESTED_MIXED_ITEMS: IMarkdownNavigatorItem[] = [
   {
     title: 'Second item',
     children: RAW_MARKDOWN_ITEM,
+  },
+];
+
+@Component({
+  template: `
+    <div>
+      Footer A Content
+    </div>
+  `,
+})
+export class FooterAComponent {}
+
+@Component({
+  template: `
+    <div>
+      Global Footer Content
+    </div>
+  `,
+})
+export class GlobalFooterComponent {}
+
+const ITEMS_WITH_FOOTERS: IMarkdownNavigatorItem[] = [
+  {
+    markdownString: `Footer A`,
+    footer: FooterAComponent,
+  },
+  {
+    markdownString: `Global Footer`,
+  },
+];
+
+const CHILDREN_URL: string = 'https://samplechildrenurl.com';
+
+const ITEMS_WITH_CHILDREN_URL: IMarkdownNavigatorItem[] = [
+  {
+    title: 'Children url',
+    childrenUrl: CHILDREN_URL,
   },
 ];
 
@@ -100,6 +152,23 @@ export const DEEPLY_NESTED_TREE: IMarkdownNavigatorItem[] = [
         ],
       },
     ],
+  },
+];
+
+export const ITEMS_AT_SAME_LEVEL_AS_MARKDOWN: IMarkdownNavigatorItem[] = [
+  {
+    title: 'A',
+    markdownString: `A markdown`,
+    children: [
+      {
+        title: 'A1',
+        markdownString: `A1 markdown`,
+      },
+    ],
+  },
+  {
+    title: 'B',
+    markdownString: 'B',
   },
 ];
 
@@ -217,6 +286,7 @@ async function validateTree(fixture: ComponentFixture<TdMarkdownNavigatorTestCom
       [labels]="labels"
       [startAt]="startAt"
       [compareWith]="compareWith"
+      [footer]="footer"
     ></td-markdown-navigator>
   `,
 })
@@ -225,14 +295,21 @@ class TdMarkdownNavigatorTestComponent {
   labels: IMarkdownNavigatorLabels;
   startAt: IMarkdownNavigatorItem;
   compareWith: IMarkdownNavigatorCompareWith;
+  footer: Type<any>;
 }
 
 describe('MarkdownNavigatorComponent', () => {
+  let httpClient: HttpClient;
+  let httpTestingController: HttpTestingController;
+
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [TdMarkdownNavigatorTestComponent],
-      imports: [NoopAnimationsModule, CovalentMarkdownNavigatorModule],
+      imports: [NoopAnimationsModule, CovalentMarkdownNavigatorModule, HttpClientTestingModule],
     }).compileComponents();
+
+    httpClient = TestBed.inject(HttpClient);
+    httpTestingController = TestBed.inject(HttpTestingController);
   }));
 
   it('should render empty state when an empty array is passed into items', async(
@@ -301,6 +378,134 @@ describe('MarkdownNavigatorComponent', () => {
     }),
   ));
 
+  it('should fetch childrenUrl and render', async(
+    inject([], async () => {
+      const itemATitle: string = 'fetched item A';
+      const itemBTitle: string = 'fetched item B';
+
+      const testData: IMarkdownNavigatorItem[] = [{ title: itemATitle }, { title: itemBTitle }];
+      const fixture: ComponentFixture<TdMarkdownNavigatorTestComponent> = TestBed.createComponent(
+        TdMarkdownNavigatorTestComponent,
+      );
+
+      fixture.componentInstance.items = ITEMS_WITH_CHILDREN_URL;
+      await wait(fixture);
+
+      const markdownNavigator: TdMarkdownNavigatorComponent = fixture.debugElement.query(
+        By.directive(TdMarkdownNavigatorComponent),
+      ).componentInstance;
+
+      expect(markdownNavigator.showMenu).toBeTruthy();
+      expect(markdownNavigator.showTdMarkdown).toBeFalsy();
+
+      getItem(fixture, 0).click();
+      const req: TestRequest = httpTestingController.expectOne(CHILDREN_URL);
+      req.flush(testData);
+
+      await wait(fixture);
+      await wait(fixture);
+
+      expect(markdownNavigator.showMenu).toBeTruthy();
+      expect(markdownNavigator.showTdMarkdown).toBeFalsy();
+      expect(getItem(fixture, 0).textContent).toContain(itemATitle);
+      expect(getItem(fixture, 1).textContent).toContain(itemBTitle);
+
+      httpTestingController.verify();
+    }),
+  ));
+
+  it('should show proper error messages', async(
+    inject([], async () => {
+      const invalidMarkdownUrl: string = 'https://invalid_markdown_url.com';
+      const invalidChildrenUrl: string = 'https://invalid_children_url.com';
+      const ITEMS_WITH_ERRORS: IMarkdownNavigatorItem[] = [
+        {
+          title: 'Invalid markdown url',
+          url: invalidMarkdownUrl,
+          children: [
+            {
+              title: 'Invalid children url',
+              childrenUrl: invalidChildrenUrl,
+            },
+          ],
+        },
+        { title: 'Invalid markdown and children url', url: invalidMarkdownUrl, childrenUrl: invalidChildrenUrl },
+      ];
+      const invalidMarkdownUrlOptions: object = { status: 404, statusText: 'Not Found' };
+      const message: string = 'Failed :(';
+      function getMarkdownLoaderError(): HTMLElement {
+        return (fixture.debugElement.query(By.css('[data-test="markdown-loader-error"]')) || {}).nativeElement;
+      }
+      function getChildrenUrlError(): HTMLElement {
+        return (fixture.debugElement.query(By.css('[data-test="children-url-error"]')) || {}).nativeElement;
+      }
+
+      const fixture: ComponentFixture<TdMarkdownNavigatorTestComponent> = TestBed.createComponent(
+        TdMarkdownNavigatorTestComponent,
+      );
+      fixture.componentInstance.items = ITEMS_WITH_ERRORS;
+
+      await wait(fixture);
+      getItem(fixture, 0).click();
+      await wait(fixture);
+
+      const invalidMarkdownUrlRequest: TestRequest = httpTestingController.expectOne(invalidMarkdownUrl);
+      invalidMarkdownUrlRequest.flush(message, invalidMarkdownUrlOptions);
+
+      await wait(fixture);
+      await wait(fixture);
+      await wait(fixture);
+
+      expect(getMarkdownLoaderError()).toBeTruthy();
+      expect(getMarkdownLoaderError().textContent).toContain('Not Found');
+
+      getItem(fixture, 0).click();
+      const invalidChildrenUrlRequest: TestRequest = httpTestingController.expectOne(invalidChildrenUrl);
+      invalidChildrenUrlRequest.flush(message, invalidMarkdownUrlOptions);
+
+      await wait(fixture);
+      await wait(fixture);
+      await wait(fixture);
+
+      expect(getMarkdownLoaderError()).toBeFalsy();
+      expect(getChildrenUrlError()).toBeTruthy();
+      goBack(fixture);
+
+      const invalidMarkdownUrlRequest2: TestRequest = httpTestingController.expectOne(invalidMarkdownUrl);
+      invalidMarkdownUrlRequest2.flush(message, invalidMarkdownUrlOptions);
+
+      await wait(fixture);
+      await wait(fixture);
+      await wait(fixture);
+
+      expect(getMarkdownLoaderError()).toBeTruthy();
+      expect(getChildrenUrlError()).toBeFalsy();
+
+      goBack(fixture);
+      getItem(fixture, 1).click();
+      await wait(fixture);
+
+      const invalidMarkdownUrlRequest3: TestRequest = httpTestingController.expectOne(invalidMarkdownUrl);
+      invalidMarkdownUrlRequest3.flush(message, invalidMarkdownUrlOptions);
+      const invalidChildrenUrlRequest2: TestRequest = httpTestingController.expectOne(invalidChildrenUrl);
+      invalidChildrenUrlRequest2.flush(message, invalidMarkdownUrlOptions);
+
+      await wait(fixture);
+      await wait(fixture);
+      await wait(fixture);
+
+      expect(getMarkdownLoaderError()).toBeTruthy();
+      expect(getChildrenUrlError()).toBeTruthy();
+
+      goBack(fixture);
+
+      expect(getMarkdownLoaderError()).toBeFalsy();
+      expect(getChildrenUrlError()).toBeFalsy();
+
+      httpTestingController.verify();
+    }),
+  ));
+
   it('should render one url item from GitHub', async(
     inject([], async () => {
       const fixture: ComponentFixture<TdMarkdownNavigatorTestComponent> = TestBed.createComponent(
@@ -347,6 +552,40 @@ describe('MarkdownNavigatorComponent', () => {
     }),
   ));
 
+  it('should use custom icons if passed in', async(
+    inject([], async () => {
+      const fixture: ComponentFixture<TdMarkdownNavigatorTestComponent> = TestBed.createComponent(
+        TdMarkdownNavigatorTestComponent,
+      );
+
+      fixture.componentInstance.items = ITEMS_WITH_CUSTOM_ICONS;
+      await wait(fixture);
+
+      const listItems: DebugElement[] = fixture.debugElement.queryAll(By.css('mat-action-list button'));
+
+      expect(listItems.length).toBe(ITEMS_WITH_CUSTOM_ICONS.length);
+      expect(listItems[0].nativeElement.textContent).toContain(ITEMS_WITH_CUSTOM_ICONS[0].icon);
+      expect(listItems[1].nativeElement.textContent).toContain(ITEMS_WITH_CUSTOM_ICONS[1].icon);
+    }),
+  ));
+
+  it('should use descriptions if passed in', async(
+    inject([], async () => {
+      const fixture: ComponentFixture<TdMarkdownNavigatorTestComponent> = TestBed.createComponent(
+        TdMarkdownNavigatorTestComponent,
+      );
+
+      fixture.componentInstance.items = ITEMS_WITH_DESCRIPTIONS;
+      await wait(fixture);
+
+      const listItems: DebugElement[] = fixture.debugElement.queryAll(By.css('mat-action-list button'));
+
+      expect(listItems.length).toBe(ITEMS_WITH_DESCRIPTIONS.length);
+      expect(listItems[0].nativeElement.textContent).toContain(ITEMS_WITH_DESCRIPTIONS[0].description);
+      expect(listItems[1].nativeElement.textContent).toContain(ITEMS_WITH_DESCRIPTIONS[1].description);
+    }),
+  ));
+
   it('should render a nested list of items', async(
     inject([], async () => {
       const fixture: ComponentFixture<TdMarkdownNavigatorTestComponent> = TestBed.createComponent(
@@ -370,6 +609,49 @@ describe('MarkdownNavigatorComponent', () => {
       expect(listItems.length).toBe(NESTED_MIXED_ITEMS.length);
       expect(listItems[0].nativeElement.textContent).toContain(NESTED_MIXED_ITEMS[0].title);
       expect(listItems[1].nativeElement.textContent).toContain(NESTED_MIXED_ITEMS[1].title);
+    }),
+  ));
+
+  it('should render list and markdown side by side', async(
+    inject([], async () => {
+      const fixture: ComponentFixture<TdMarkdownNavigatorTestComponent> = TestBed.createComponent(
+        TdMarkdownNavigatorTestComponent,
+      );
+
+      fixture.componentInstance.items = ITEMS_AT_SAME_LEVEL_AS_MARKDOWN;
+      await wait(fixture);
+
+      const markdownNavigator: TdMarkdownNavigatorComponent = fixture.debugElement.query(
+        By.directive(TdMarkdownNavigatorComponent),
+      ).componentInstance;
+      const listItems: DebugElement[] = fixture.debugElement.queryAll(By.css('mat-action-list button'));
+
+      expect(markdownNavigator.showTdMarkdown).toBeFalsy();
+      expect(markdownNavigator.showTdMarkdownLoader).toBeFalsy();
+      expect(listItems.length).toBe(ITEMS_AT_SAME_LEVEL_AS_MARKDOWN.length);
+      expect(listItems[0].nativeElement.textContent).toContain(ITEMS_AT_SAME_LEVEL_AS_MARKDOWN[0].title);
+      expect(listItems[1].nativeElement.textContent).toContain(ITEMS_AT_SAME_LEVEL_AS_MARKDOWN[1].title);
+      getItem(fixture, 0).click();
+      await wait(fixture);
+
+      expect(markdownNavigator.showMenu).toBeTruthy();
+      expect(markdownNavigator.showTdMarkdown).toBeTruthy();
+      expect(getTitle(fixture)).toContain(ITEMS_AT_SAME_LEVEL_AS_MARKDOWN[0].title);
+      expect(getMarkdown(fixture)).toContain(ITEMS_AT_SAME_LEVEL_AS_MARKDOWN[0].markdownString);
+
+      getItem(fixture, 0).click();
+      await wait(fixture);
+
+      expect(markdownNavigator.showMenu).toBeFalsy();
+      expect(markdownNavigator.showTdMarkdown).toBeTruthy();
+      expect(getTitle(fixture)).toContain(ITEMS_AT_SAME_LEVEL_AS_MARKDOWN[0].children[0].title);
+      expect(getMarkdown(fixture)).toContain(ITEMS_AT_SAME_LEVEL_AS_MARKDOWN[0].children[0].markdownString);
+      goBack(fixture);
+
+      expect(markdownNavigator.showMenu).toBeTruthy();
+      expect(markdownNavigator.showTdMarkdown).toBeTruthy();
+      expect(getTitle(fixture)).toContain(ITEMS_AT_SAME_LEVEL_AS_MARKDOWN[0].title);
+      expect(getMarkdown(fixture)).toContain(ITEMS_AT_SAME_LEVEL_AS_MARKDOWN[0].markdownString);
     }),
   ));
 
@@ -529,7 +811,38 @@ describe('MarkdownNavigatorComponent', () => {
       fixture.componentInstance.items = NESTED_MIXED_ITEMS;
       fixture.componentInstance.startAt = NESTED_MIXED_ITEMS[1];
       await wait(fixture);
-      expect(getMarkdown(fixture)).toContain('Heading');
+      expect(getTitle(fixture)).toContain(NESTED_MIXED_ITEMS[1].title);
+    }),
+  ));
+
+  it('should be able to render a custom component as a footer', async(
+    inject([], async () => {
+      const fixture: ComponentFixture<TdMarkdownNavigatorTestComponent> = TestBed.createComponent(
+        TdMarkdownNavigatorTestComponent,
+      );
+
+      fixture.componentInstance.items = ITEMS_WITH_FOOTERS;
+      fixture.componentInstance.footer = GlobalFooterComponent;
+
+      await wait(fixture);
+
+      expect(fixture.nativeElement.textContent).toContain('Global Footer Content');
+
+      getItem(fixture, 0).click();
+
+      await wait(fixture);
+
+      expect(fixture.nativeElement.textContent).toContain('Footer A Content');
+
+      goBack(fixture);
+
+      await wait(fixture);
+
+      getItem(fixture, 1).click();
+
+      await wait(fixture);
+
+      expect(fixture.nativeElement.textContent).toContain('Global Footer Content');
     }),
   ));
 });