+++ src/platform/markdown-navigator/markdown-navigator.component.ts
@@ -8,16 +8,28 @@ import {
   ElementRef,
   ChangeDetectorRef,
   ChangeDetectionStrategy,
+  Type,
+  Output,
+  EventEmitter,
+  SecurityContext,
 } from '@angular/core';
 import { removeLeadingHash, isAnchorLink, TdMarkdownLoaderService } from '@covalent/markdown';
+import { ITdFlavoredMarkdownButtonClickEvent } from '@covalent/flavored-markdown';
+import { DomSanitizer } from '@angular/platform-browser';
+import { HttpClient } from '@angular/common/http';
 
 export interface IMarkdownNavigatorItem {
+  id?: string;
   title?: string;
   url?: string;
   httpOptions?: object;
   markdownString?: string; // raw markdown
   anchor?: string;
   children?: IMarkdownNavigatorItem[];
+  childrenUrl?: string;
+  description?: string;
+  icon?: string;
+  footer?: Type<any>;
 }
 
 export interface IMarkdownNavigatorLabels {
@@ -60,6 +72,9 @@ function isMarkdownHref(anchor: HTMLAnchorElement): boolean {
   return !isAnchorLink(anchor) && anchor.pathname.endsWith('.md');
 }
 function defaultCompareWith(o1: IMarkdownNavigatorItem, o2: IMarkdownNavigatorItem): boolean {
+  if (o1.id && o2.id) {
+    return o1.id === o2.id;
+  }
   return o1 === o2;
 }
 
@@ -111,12 +126,21 @@ export class TdMarkdownNavigatorComponent implements OnChanges {
   @Input() startAt: IMarkdownNavigatorItem;
 
   /**
+   * footer?: Type<any>
+   *
+   * Component to be displayed in footer
+   */
+  @Input() footer: Type<any>;
+
+  /**
    * compareWith?: IMarkdownNavigatorCompareWith
    *
    * Function used to find startAt item
    * Defaults to comparison by strict equality (===)
    */
   @Input() compareWith: IMarkdownNavigatorCompareWith;
+
+  @Output() buttonClicked: EventEmitter<ITdFlavoredMarkdownButtonClickEvent> = new EventEmitter();
 
   @ViewChild('markdownWrapper') markdownWrapper: ElementRef;
 
@@ -126,9 +150,14 @@ export class TdMarkdownNavigatorComponent implements OnChanges {
 
   loading: boolean = false;
 
+  markdownLoaderError: string;
+  childrenUrlError: string;
+
   constructor(
     private _markdownUrlLoaderService: TdMarkdownLoaderService,
     private _changeDetectorRef: ChangeDetectorRef,
+    private _sanitizer: DomSanitizer,
+    private _http: HttpClient,
   ) {}
 
   @HostListener('click', ['$event'])
@@ -168,6 +197,13 @@ export class TdMarkdownNavigatorComponent implements OnChanges {
       return this.currentMarkdownItem.url;
     }
     return undefined;
+  }
+
+  get footerComponent(): any {
+    if (this.currentMarkdownItem && this.currentMarkdownItem.footer) {
+      return this.currentMarkdownItem.footer;
+    }
+    return this.footer;
   }
   get httpOptions(): object {
     if (this.currentMarkdownItem) {
@@ -222,9 +258,19 @@ export class TdMarkdownNavigatorComponent implements OnChanges {
     }
   }
 
+  hasChildrenOrChildrenUrl(item: IMarkdownNavigatorItem): boolean {
+    return (item.children && item.children.length > 0) || !!item.childrenUrl;
+  }
+  clearErrors(): void {
+    this.markdownLoaderError = undefined;
+    this.childrenUrlError = undefined;
+  }
+
   reset(): void {
+    this.loading = false;
+    this.clearErrors();
     // if single item and no children
-    if (this.items && this.items.length === 1 && (!this.items[0].children || this.items[0].children.length === 0)) {
+    if (this.items && this.items.length === 1 && !this.hasChildrenOrChildrenUrl(this.items[0])) {
       this.currentMenuItems = [];
       this.currentMarkdownItem = this.items[0];
     } else {
@@ -236,18 +282,13 @@ export class TdMarkdownNavigatorComponent implements OnChanges {
   }
 
   goBack(): void {
+    this.loading = false;
+    this.clearErrors();
     if (this.historyStack.length > 1) {
       const parent: IMarkdownNavigatorItem = this.historyStack[this.historyStack.length - 2];
-      if (parent.children && parent.children.length > 0) {
-        // if parent has children, show menu
-        this.currentMenuItems = parent.children;
-        this.currentMarkdownItem = undefined;
-      } else {
-        // else just render markdown
-        this.currentMenuItems = [];
-        this.currentMarkdownItem = parent;
-      }
+      this.currentMarkdownItem = parent;
       this.historyStack = this.historyStack.slice(0, -1);
+      this.setChildrenAsCurrentMenuItems(parent);
     } else {
       // one level down just go to root
       this.reset();
@@ -256,27 +297,47 @@ export class TdMarkdownNavigatorComponent implements OnChanges {
   }
 
   handleItemSelected(item: IMarkdownNavigatorItem): void {
+    this.clearErrors();
+    this.currentMarkdownItem = item;
     this.historyStack = [...this.historyStack, item];
-    if (
-      item.children &&
-      item.children.length === 1 &&
-      (!item.children[0].children || item.children[0].children.length === 0)
-    ) {
-      // clicked on item with one child that has no children
-      // don't show menu
-      this.currentMenuItems = [];
-      // render markdown
-      this.currentMarkdownItem = item.children[0];
-    } else if (item.children && item.children.length > 0) {
-      // has children, go inside
-      this.currentMenuItems = item.children;
-    } else {
-      // don't show menu
-      this.currentMenuItems = [];
-      // render markdown
-      this.currentMarkdownItem = item;
-    }
+    this.setChildrenAsCurrentMenuItems(item);
     this._changeDetectorRef.markForCheck();
+  }
+
+  async setChildrenAsCurrentMenuItems(item: IMarkdownNavigatorItem): Promise<void> {
+    this.currentMenuItems = [];
+    this.loading = true;
+    this._changeDetectorRef.markForCheck();
+
+    const stackSnapshot: IMarkdownNavigatorItem[] = this.historyStack;
+    let children: IMarkdownNavigatorItem[] = [];
+    if (item.children) {
+      children = item.children;
+    } else if (item.childrenUrl) {
+      children = await this.loadChildrenUrl(item);
+    }
+    const newStackSnapshot: IMarkdownNavigatorItem[] = this.historyStack;
+    if (
+      stackSnapshot.length === newStackSnapshot.length &&
+      stackSnapshot.every((stackItem: IMarkdownNavigatorItem, index: number) => stackItem === newStackSnapshot[index])
+    ) {
+      this.currentMenuItems = children;
+    }
+
+    this.loading = false;
+    this._changeDetectorRef.markForCheck();
+  }
+
+  async loadChildrenUrl(item: IMarkdownNavigatorItem): Promise<IMarkdownNavigatorItem[]> {
+    const sanitizedUrl: string = this._sanitizer.sanitize(SecurityContext.URL, item.childrenUrl);
+    try {
+      return await this._http
+        .get<IMarkdownNavigatorItem[]>(sanitizedUrl, { ...item.httpOptions })
+        .toPromise();
+    } catch (error) {
+      this.handleChildrenUrlError(error);
+      return [];
+    }
   }
 
   getTitle(item: IMarkdownNavigatorItem): string {
@@ -289,6 +350,21 @@ export class TdMarkdownNavigatorComponent implements OnChanges {
         ''
       ).trim();
     }
+  }
+
+  getIcon(item: IMarkdownNavigatorItem): string {
+    if (item) {
+      return item.icon || 'subject';
+    }
+  }
+
+  handleChildrenUrlError(error: Error): void {
+    this.childrenUrlError = error.message;
+    this._changeDetectorRef.markForCheck();
+  }
+  handleMarkdownLoaderError(error: Error): void {
+    this.markdownLoaderError = error.message;
+    this._changeDetectorRef.markForCheck();
   }
 
   private _jumpTo(item: IMarkdownNavigatorItem): void {