+++ src/platform/markdown-navigator/markdown-navigator.module.ts
@@ -8,8 +8,8 @@ import { MatListModule } from '@angular/material/list';
 import { MatIconModule } from '@angular/material/icon';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
 import { CovalentFlavoredMarkdownModule } from '@covalent/flavored-markdown';
-import { MatToolbarModule } from '@angular/material/toolbar';
 import { CovalentDialogsModule } from '@covalent/core/dialogs';
+import { CovalentMessageModule } from '@covalent/core/message';
 import { TdMarkdownNavigatorWindowDirective } from './markdown-navigator-window-directive/markdown-navigator-window.directive';
 import { TdMarkdownNavigatorWindowService } from './markdown-navigator-window-service/markdown-navigator-window.service';
 
@@ -23,8 +23,7 @@ import { TdMarkdownNavigatorWindowService } from './markdown-navigator-window-se
     MatListModule,
     MatIconModule,
     MatProgressBarModule,
-    MatToolbarModule,
-
+    CovalentMessageModule,
     CovalentFlavoredMarkdownModule,
     CovalentDialogsModule,
   ],